+++ Makefile
@@ -19,7 +19,7 @@ LIBFT_DIR = libft/
 SRCS = main.c minishell.c \
 parser.c \
 executer.c \
-signal.c signal_handlers.c \
+signals/signal.c signals/signal_handlers.c \
 exit.c
 SRCS_DIR = srcs/
 OBJS = $(addprefix $(OBJS_DIR), $(SRCS:.c=.o))
@@ -39,6 +39,7 @@ $(OBJS_DIR)%.o : $(SRCS_DIR)%.c
 
 $(OBJS_DIR):
 	mkdir -p objs
+	mkdir -p objs/signals/
 
 clean:
 	$(RM) $(OBJS)
